Add PowerShell module installation
As part of the PowerShell configuration we need to install a couple of
modules that helps improve our experience on the shell. However, as the
PowerShell Desired State Configuration creates a dedicated PowerShell
session for each configuration. It uses a $Env:PSModulePath that limits
automatic module loading to only the path where the package was
extracted. Therefore, we built the PSModulePath manually.